backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
pydantic
google-generativeai
python-dotenv
selenium
webdriver-manager
fpdf2
bleach
langdetect
matplotlib